Avoid using a keyword as a measurement name in an example.
If the reader attempts to use a measurement with a name of "measurement"
in a query then they will likely omit the required double-quoting of
measurement where it is used as a measurement name and then receive a
confusing report of syntax error as a result.

With this change, the measurement name used in the example will not
require quoting see https://github.com/influxdata/influxdb/issues/6785
if subsequently used.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>